backend/requirements.txt
# Scanner Transcription Backend
websockets>=12.0
openai-whisper>=20231117

# Optional: faster-whisper for better performance (requires CUDA)
# faster-whisper>=0.10.0
